// packages/datadog-api-client-v1/models/LogsListRequestTime.d.ts
/**
 * Unless explicitly stated otherwise all files in this repository are licensed under the Apache-2.0 License.
 * This product includes software developed at Datadog (https://www.datadoghq.com/).
 * Copyright 2020-Present Datadog, Inc.
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
/**
* Timeframe to retrieve the log from.
*/
export declare class LogsListRequestTime {
    /**
    * Minimum timestamp for requested logs.
    */
    'from': Date;
    /**
    * Timezone can be specified both as an offset (e.g. \"UTC+03:00\") or a regional zone (e.g. \"Europe/Paris\").
    */
    'timezone'?: string;
    /**
    * Maximum timestamp for requested logs.
    */
    'to': Date;
    static readonly discriminator: string | undefined;
    static readonly attributeTypeMap: {
        [key: string]: {
            baseName: string;
            type: string;
            format: string;
        };
    };
    static getAttributeTypeMap(): {
        [key: string]: {
            baseName: string;
            type: string;
            format: string;
        };
    };
    static deserialize(data: {
        [key: string]: any;
    }): LogsListRequestTime;
    static serialize(data: LogsListRequestTime): {
        [key: string]: any;
    };
    constructor();
}